+++ src/tea/process/win_process.py
@@ -14,6 +14,7 @@ import win32file  # @UnresolvedImport
 import win32event  # @UnresolvedImport
 import win32process  # @UnresolvedImport
 import win32security  # @UnresolvedImport
+from tea.process.base import Process
 
 
 def create_file(filename, mode='rw'):
@@ -35,7 +36,7 @@ def create_file(filename, mode='rw'):
     return handle
 
 
-class Process(object):
+class WinProcess(Process):
     def __init__(self, command, arguments=None, environment=None, redirect_output=True):
         # Parse and generate the command line
         self._commandline       = self._get_cmd(command, [] if arguments is None else arguments)
@@ -121,17 +122,18 @@ class Process(object):
                                        self._currentDirectory, self._startupinfo)
 
     def kill(self):
-        return Process.Kill(self.pid)
+        return self.Kill(self.pid)
 
     def wait(self, timeout=None):
         if timeout is None:
-            while self.is_running():
+            while self.is_running:
                 win32api.Sleep(1000)
         else:
             if win32event.WaitForSingleObject(self._hProcess, timeout * 1000) != win32event.WAIT_OBJECT_0:
                 return False
         return True
 
+    @property
     def is_running(self):
         if self._hProcess is None or self._exit_code is not None:
             return False
@@ -141,11 +143,13 @@ class Process(object):
             return False
         return True
 
-    def _get_pid(self):
+    @property
+    def pid(self):
         return self._dwProcessId
 
-    def _get_exit_code(self):
-        if self.is_running():
+    @property
+    def exit_code(self):
+        if self.is_running:
             return None
         return self._exit_code
 
@@ -169,7 +173,7 @@ class Process(object):
 #    def _get_stdout(self):
 #        return self._stdout.read()
 #        #if self._stdout is None:
-#        #    if self.is_running():
+#        #    if self.is_running:
 #        #        return ''
 #        #    else:
 #        #        win32file.CloseHandle(self._stdout_write)
@@ -183,7 +187,7 @@ class Process(object):
 #    def _get_stderr(self):
 #        return self._stderr.read()
 #        #if self._stderr is None:
-#        #    if self.is_running():
+#        #    if self.is_running:
 #        #        return ''
 #        #    else:
 #        #        win32file.CloseHandle(self._stderr_write)
@@ -194,23 +198,8 @@ class Process(object):
 #        #            self._stderr = ''
 #        #return self._stderr
 
-    pid       = property(_get_pid)
-    exit_code = property(_get_exit_code)
-    #stdin     = property(_get_stdin)
-    #stdout    = property(_get_stdout)
-    #stderr    = property(_get_stderr)
-
-    @staticmethod
-    def GetProcesses(sort_by_name=True, cmdline=False):
-        '''Retrieves a list of processes sorted by name.
-
-        @type  sort_by_name: boolean
-        @param sort_by_name: Sort the list by name or by PID
-        @type  cmdline: boolean
-        @param cmdline: Add process command line to output
-        @rtype:  list[tuple(string, string)]
-        @return: List of process PID, process name tuples
-        '''
+    @classmethod
+    def GetProcesses(cls, sort_by_name=True, cmdline=False):
         if cmdline:
             processes = [(p.pid, p.name, p.cmdline) for p in psutil.get_process_list()]
         else:
@@ -220,33 +209,24 @@ class Process(object):
         else:
             return sorted(processes, lambda t1, t2: cmp(t1[0], t2[0]) or cmp(t1[1], t2[1]))
 
-    @staticmethod
-    def Find(name, arg=None):
-        '''Find process by name or by argument in command line if arg param is available'''
+    @classmethod
+    def Find(cls, name, arg=None):
         if arg is None:
-            for pid, process in Process.GetProcesses():
+            for pid, process in cls.GetProcesses():
                 if process.lower().find(name.lower()) != -1:
                     return process, pid
         else:
-            for pid, process, cmdline in Process.GetProcesses(cmdline=True):
+            for pid, process, cmdline in cls.GetProcesses(cmdline=True):
                 if process.lower().find(name.lower()) != -1:
                     if cmdline is not None and cmdline.lower().find(arg.lower()) != -1:
                         return process, pid
         return None
 
-    @staticmethod
-    def Kill(pid=None, process=None):
-        '''Kills a process by process PID or
-        kills a process started by process module.
-
-        @type pid: int
-        @param pid: Process ID of the process to kill
-        @type  process: ShellProcess
-        @param process: Process started by process module
-        '''
+    @classmethod
+    def Kill(cls, pid=None, process=None):
         if process is not None:
             pid = process.pid
-        process = Process(os.path.join(os.environ['windir'], 'system32', 'taskkill.exe'), ['/PID', str(pid), '/F', '/T'])
+        process = cls(os.path.join(os.environ['windir'], 'system32', 'taskkill.exe'), ['/PID', str(pid), '/F', '/T'])
         process.start()
         process.wait()
         return process.exit_code == 0